iot-simulator/src/real_device.py
"""
Real ESP32 Device Interface - For when you have actual hardware
This implements the same DeviceInterface as the simulator
"""

import asyncio
import json
import ssl
from typing import Dict, Any, Optional
from datetime import datetime
import paho.mqtt.client as mqtt
from pathlib import Path

from .device_interface import (
    DeviceInterface, 
    SensorReading, 
    DeviceDiagnostics, 
    WaterQualityData
)


class RealESP32Device(DeviceInterface):
    """Real ESP32 device implementation - use when hardware is available"""
    
    def __init__(self, device_id: str, location: Dict[str, float], 
                 certificate_path: str, config: Dict[str, Any]):
        super().__init__(device_id, location)
        self.certificate_path = Path(certificate_path)
        self.config = config
        self.mqtt_client = None
        self.is_connected = False
        
        # Certificate files (generated during device provisioning)
        self.cert_file = self.certificate_path / f"{device_id}-certificate.pem.crt"
        self.private_key = self.certificate_path / f"{device_id}-private.pem.key"
        self.ca_file = self.certificate_path / "AmazonRootCA1.pem"
    
    async def initialize(self) -> bool:
        """Initialize real ESP32 device connection"""
        try:
            print(f"🔄 Connecting real device {self.device_id}...")
            
            # Verify certificate files exist
            if not all(f.exists() for f in [self.cert_file, self.private_key, self.ca_file]):
                print(f"❌ Certificate files missing for {self.device_id}")
                return False
            
            # Set up MQTT client with certificates
            self.mqtt_client = mqtt.Client(
                client_id=f"esp32-{self.device_id}-{int(datetime.utcnow().timestamp())}"
            )
            
            # Configure TLS
            context = ssl.create_default_context(ssl.Purpose.SERVER_AUTH)
            context.check_hostname = False
            context.verify_mode = ssl.CERT_REQUIRED
            context.load_verify_locations(str(self.ca_file))
            context.load_cert_chain(str(self.cert_file), str(self.private_key))
            
            self.mqtt_client.tls_set_context(context)
            
            # Set up callbacks
            self.mqtt_client.on_connect = self._on_connect
            self.mqtt_client.on_disconnect = self._on_disconnect
            self.mqtt_client.on_publish = self._on_publish
            
            # Connect to AWS IoT Core
            endpoint = self.config['aws']['iot_endpoint']
            port = self.config['mqtt']['port']
            
            self.mqtt_client.connect(endpoint, port, self.config['mqtt']['keepalive'])
            self.mqtt_client.loop_start()
            
            # Wait for connection
            for _ in range(10):  # 10 second timeout
                if self.is_connected:
                    break
                await asyncio.sleep(1)
            
            if self.is_connected:
                self.is_online = True
                print(f"✅ Real device {self.device_id} connected successfully")
                return True
            else:
                print(f"❌ Failed to connect real device {self.device_id}")
                return False
                
        except Exception as e:
            print(f"❌ Error initializing real device {self.device_id}: {e}")
            return False
    
    async def read_sensors(self) -> SensorReading:
        """
        Read sensors from real ESP32 hardware
        
        In a real implementation, this would:
        1. Read from I2C/SPI sensor interfaces
        2. Apply calibration factors
        3. Handle sensor errors
        """
        
        # TODO: Replace with actual sensor reading code
        # This is a placeholder that would be replaced with:
        # - pH sensor reading via analog input
        # - Turbidity sensor via I2C
        # - TDS sensor via analog input
        # - Temperature/humidity via DHT22 or similar
        
        raise NotImplementedError(
            "Real sensor reading not implemented. "
            "This requires ESP32 firmware with sensor drivers."
        )
        
        # Example of what real implementation would look like:
        """
        try:
            # Read pH sensor (analog input with calibration)
            ph_raw = analogRead(pH_PIN)
            ph_value = calibrate_ph(ph_raw)
            
            # Read turbidity sensor (I2C)
            turbidity_value = turbidity_sensor.read()
            
            # Read TDS sensor (analog input)
            tds_raw = analogRead(TDS_PIN)
            tds_value = calibrate_tds(tds_raw, temperature)
            
            # Read temperature/humidity (DHT22)
            temp_value, humidity_value = dht22.read()
            
            return SensorReading(
                pH=ph_value,
                turbidity=turbidity_value,
                tds=int(tds_value),
                temperature=temp_value,
                humidity=humidity_value
            )
        except Exception as e:
            # Handle sensor errors
            raise SensorError(f"Failed to read sensors: {e}")
        """
    
    async def get_diagnostics(self) -> DeviceDiagnostics:
        """Get real device diagnostics"""
        
        # TODO: Replace with actual hardware diagnostics
        # This would read:
        # - Battery voltage via ADC
        # - WiFi signal strength
        # - Sensor status from hardware
        # - System uptime
        
        raise NotImplementedError(
            "Real diagnostics not implemented. "
            "This requires ESP32 firmware with hardware monitoring."
        )
        
        # Example implementation:
        """
        try:
            battery_voltage = analogRead(BATTERY_PIN) * VOLTAGE_DIVIDER_RATIO
            battery_percentage = voltage_to_percentage(battery_voltage)
            
            wifi_rssi = WiFi.RSSI()
            
            sensor_status = check_sensor_health()
            
            uptime_ms = millis()
            
            return DeviceDiagnostics(
                batteryLevel=battery_percentage,
                signalStrength=wifi_rssi,
                sensorStatus=sensor_status,
                firmwareVersion=FIRMWARE_VERSION,
                uptime=uptime_ms // 1000
            )
        except Exception as e:
            raise DiagnosticsError(f"Failed to get diagnostics: {e}")
        """
    
    async def send_data(self, data: WaterQualityData) -> bool:
        """Send data via MQTT (same as simulator)"""
        try:
            if not self.is_connected:
                print(f"❌ Device {self.device_id} not connected")
                return False
            
            topic = self.config['mqtt']['topic_template'].format(deviceId=self.device_id)
            payload = data.model_dump_json()
            
            result = self.mqtt_client.publish(
                topic=topic,
                payload=payload,
                qos=self.config['mqtt']['qos']
            )
            
            if result.rc == mqtt.MQTT_ERR_SUCCESS:
                print(f"📤 {self.device_id}: Sent real sensor data")
                return True
            else:
                print(f"❌ Failed to publish data for {self.device_id}: {result.rc}")
                return False
                
        except Exception as e:
            print(f"❌ Error sending data for {self.device_id}: {e}")
            return False
    
    def _on_connect(self, client, userdata, flags, rc):
        """MQTT connection callback"""
        if rc == 0:
            self.is_connected = True
            print(f"🔗 MQTT connected for {self.device_id}")
        else:
            print(f"❌ MQTT connection failed for {self.device_id}: {rc}")
    
    def _on_disconnect(self, client, userdata, rc):
        """MQTT disconnection callback"""
        self.is_connected = False
        print(f"🔌 MQTT disconnected for {self.device_id}: {rc}")
    
    def _on_publish(self, client, userdata, mid):
        """MQTT publish callback"""
        pass  # Could add publish confirmation logging here
    
    async def shutdown(self):
        """Shutdown real device connection"""
        print(f"🔌 Shutting down real device {self.device_id}")
        if self.mqtt_client:
            self.mqtt_client.loop_stop()
            self.mqtt_client.disconnect()
        self.is_online = False
        self.is_connected = False


class RealDeviceFactory:
    """Factory to create real devices from configuration"""
    
    @staticmethod
    def create_from_config(device_config: Dict[str, Any], 
                          aws_config: Dict[str, Any]) -> RealESP32Device:
        """Create a real device from JSON configuration"""
        
        device_id = device_config['deviceId']
        location = device_config['location']
        certificate_path = aws_config['aws']['certificate_path']
        
        return RealESP32Device(
            device_id=device_id,
            location=location,
            certificate_path=certificate_path,
            config=aws_config
        )


# Utility functions for ESP32 firmware (would be implemented in C++)
"""
// ESP32 firmware functions that would be called by the Python interface

float calibrate_ph(int raw_value) {
    // Apply calibration curve for pH sensor
    // This would be determined during sensor calibration
    float voltage = raw_value * (3.3 / 4095.0);
    return 7.0 + (voltage - 1.65) * 3.5;  // Example calibration
}

float calibrate_tds(int raw_value, float temperature) {
    // Apply temperature compensation for TDS
    float voltage = raw_value * (3.3 / 4095.0);
    float tds_value = voltage * 1000;  // Example conversion
    
    // Temperature compensation
    float compensation_coefficient = 1.0 + 0.02 * (temperature - 25.0);
    return tds_value / compensation_coefficient;
}

String check_sensor_health() {
    // Check if sensors are responding correctly
    if (ph_sensor_ok && turbidity_sensor_ok && tds_sensor_ok) {
        return "normal";
    } else {
        return "fault";
    }
}
"""