iot-simulator/requirements.txt
# AquaChain IoT Simulator Dependencies
boto3>=1.34.0
paho-mqtt>=1.6.1
pydantic>=2.5.0
python-dotenv>=1.0.0
schedule>=1.2.0
colorama>=0.4.6
rich>=13.7.0
asyncio-mqtt>=0.16.1
faker>=20.1.0